Move rpc helpers to rpc file
